requirements.txt
torch
pillow
fastapi
keras
tensorflow==2.3.0
torchvision